build: replace custom variables with CMake's default
CMake has the ability to configure the default output location of
libraries and executables.  Use those variables rather than trying to
construct a custom mechanism for that.  Doing that will allow us to
simplify the install rules.  Additionally, this will make it easier to
support proper Swift builds with CMake 3.15+.